WIP: Refactor operator and vector linear algebra, adding ComplexVector, ComplexOperator, and the parallel (unassembled) ParOperator and ComplexParOperator
This is a first step towards removing a lot of the dependency on PETSc for complex-valued linear algebra and streamlining the interface to support partial assembly. Includes the necessary changes for SpaceOperator, CurlCurlOperator, and LaplaceOperator in the matrix and vector construction. Also updates operator construction from driver classes. Next up, linear solvers.