Introduce Solve Estimate Mark Refine adaptation for conformal meshes
- Primary method added to BaseSolver that implements a SolveEstimateMarkRefine loop.
- Marking is performed using a parallelized Dorfler marking scheme, aka bulk chasing, wherein a minimum
number of elements are marked for refinement in order to refine a fraction of the total error. This
marking scheme is particularly efficient in combination with higher order schemes wherein the error can
be highly concentrated in a tiny fraction of the total number of elements.
- The loop has capability for handling nonconformal meshes, but in this commit that is disabled
pending additional work to address issues within MFEM. As such, the options within the Adaptation
portion of the config file are present, but the options to specify them have been commented out.
- This also introduces an ability to perform rebalancing of conformal meshes through a gather/scatter
model wherein the mesh is serialized to the root node, before being repartitioned.
- Transient solves are not supported, as the ability to coarsen (only possible with nonconformal adaptation
presently) is not present.