Introduce flux error estimation based on a global ZZ solve.
- Introduces the FluxProjector class, which is used for computing a "coarse flux" in
a vector L2 space, then performing a global mass matrix inversion to compute a
"smooth" flux that better represents the underlying physical quantity. This is achieved
through a multigrid hierarchy.
- Introduces the CurlFluxErrorEstimator and GradFluxErrorEstimator classes, which utilize the
FluxProjector class to compute elemental error estimates comparing the coarse numerical flux
from the solution against a smoothed flux from a global mass matrix inversion.
- Introduces the CurlFluxErrorEstimator class which utilizes a flux projector to compute a
smooth flux in an ND space. Provides methods for real valued vectors, and complex vectors.
The flux is assumed to take the form μ⁻¹∇ × V, where V is either A or E depending on the solver mode.
- Introduces the GradFluxErrorEstimator class which utilizes a flux projector to compute a
smooth flux in an H1^d space. The flux is assumed to be of the form ϵ ∇ ϕ
- Introduce the CurlFluxCoefficient and GradFluxCoefficient classes which compute the numerical Curl
and Grad fluxes given a solution.