Regenerate rings data for pi/6 around (0,0,1)
- All the non-probe data is very close to the original values, modulo a few sf
- The probe data requires a transformation of the rebaselined data:
```
julia> Ref(AngleAxis(pi/6, 0,0,1)) .* [[-1.934600007e+00, +1.332005966e+00, +1.007689041e+02],
				       [-2.056769746e-03, +3.073509047e-03, +1.007227123e+01]]
2-element Vector{StaticArraysCore.SVector{3, Float64}}:
 [-2.3414157352235527, 0.18625100104843154, 100.7689041]
 [-0.0033179693732712672, 0.0016333520404633007, 10.07227123]
```
which compares to
```
            i,                  B_x[1],                  B_y[1],                  B_z[1]
 1.000000e+00,        -2.471968830e+00,        +3.021551217e-01,        +1.012769086e+02
 2.000000e+00,        -7.789873394e-03,        +1.048294735e-03,        +1.006816729e+01
```
the additional discrepancy can be partially attributed to breaking of rotational symmetry, as the axis
does not pass through a vertex of the bounding box of the domain. Better agreement can be found with
pi/3 around (1,1,1), but (0,0,1) is chosen for ease of understanding in the examples.

Note: the data was generated on a mac M1 with a tighter tolerance of 1e-9, as this was required to reduce discrepancy
in the values of the second probe and to agree with CI.