Add Quick Start tutorial
This commit adds a "Quick Start" tutorial that walks new users through
their first simulation by providing a different view on the "two sphere"
example. This tutorial is partially run and populated when the
documentation is generated, so I updated the `docs.yml` GitHub action.

This commit also adds comments to the `examples/spheres/mesh.jl` file,
with the hope to be more accessible to users with no prior gmsh
experience (complementing the tutorial, which does not cover meshing).