Update MFEM
I wanted to test out two PRs to MFEM and found that managing all the
diffs quickly got too impractical (the two PRs are against MFEM's
`master`, so I have to be a little careful with the diffs).

Updating MFEM to `master` removes part of the complexity and allows me
to focus on the other things.

(I also wanted to try the workflow for updating MFEM.)

I removed one of the patches because it got merged:
https://github.com/mfem/mfem/pull/4923