Extend infra for Serial/Parallel/GPU unit tests
This commit extends Palace infrastructure and documentation to enable
running different type of unit tests.

Now, `unit-test` will execute different tests depending on how it is
run:
- tests marked as `Serial` when there is only one MPI process
- tests marked as `Parallel` when there are multiple processes
- tests marked as `GPU` when the `--device` is `cuda` or `hip`

To enable this, tests have to be explicitly marked with the relevant
tag. I added an example and a tutorial to show how this is done, and I
expanded upon the documentation on testing.

Explicitly marking tests as `Serial` opens to door to running multiple
`Serial` tests concurrently (although, we might need to be careful if
tests perform file system operations or use other share resources to
avoid race conditions). `Parallel` tests are already run with as many
cores as available. (Currently on `main`, `Serial` tests are being run
with MPI, which means that N copies of the same identical test are being
run at the same time)

The implementation takes inspiration from MFEM's unit testing approach,
but with a significant difference. While MFEM defines different
executables for serial, parallel (`punit`), and GPU tests (`gunit` and
`pgunit`), the approach implemented keeps everything unified and allows
one to reuse the same test for different purposes simply by adding the
relevant execution tag.