Automatically choose backend when device is GPU
This commit adds a small quality of life improvement where if tests are
run with GPU and no backend is specified, the default for the current
GPU is used.

This way, tests can be run with
```
unit-tests --device cuda
```
instead of
```
unit-tests --device cuda --backend /gpu/cuda
```

Specifying a backend overrides the default.