Enable documenter/deploy workflow
Documenter.jl, which we use to render to documentation, can provide a
useful feature: posting back a status with the link of the rendered
documentation. This is particularly useful with PRs: now, after the
documentation is generated, you can click on the `documenter/deploy`
green workflow and see the documentation.

This is a feature that available by default in Documenter, but it was
not working in Palace. To find what I needed to do, I searched for
`documenter/deploy` in the Documenter codebase and found one match:
https://github.com/JuliaDocs/Documenter.jl/blob/5d2d03af73d54b138ec478280c9ef85bcf0b3ffb/src/deployconfig.jl#L515

Here, I learned that the function was posting a GitHub status in a
try-catch block. Then, I cross referenced this with the documentation
and found that we needed to enable this additional write permission.