Automatically set CMAKE_CUDA_ARCHITECTURES
Starting CMake 3.24, CMake now accepts `native` as argument for
CMAKE_CUDA_ARCHITECTURES. For CUDA 11.5+, this is passed down to `nvcc`.
For previous versions, it automatically detects the native version.

Unfortunately, we cannot use this system because magma does not accept
`sm_native` as as a valid architecture.

So, we query the output of nvidia-smi, following what hypre does as well.