Update hypre, MFEM, PETSC, and SELPc
The tagged commit for `hypre` is incompatible with CUDA 13, so that
Palace does not compile with the latest version of the CUDA toolkit.

This commit updates a few packages to their latest versions and adds a
new dependency on Umpire. Umpire is strongly recommended for hypre 3
with CUDA.

Similar to magma, umpire is automatically added to the build when
buidling with CUDA or HIP.