Force hypre to compile with C++17
PR #519 is almost ready, but it uncovered a couple of possible problems.
For this reason, we might want to be cautious with it.

One of the key benefits of PR #519 is that it brings compatibility with
CUDA 13 and with CUDA 12.9 + GCC 14. This PR does not fix compatibility
with CUDA 13, but it does for CUDA 12.9 + GCC 14.