Use ctest, parallelize tests, report coverage
This commit introduces a few improvements to the test framework:

* Support for `ctest`

Tests can now be run with:
```
ctest -j $(nproc)
```
This will automatically take care of running all the tests, including
the MPI tests (with two process) and the GPU tests (if applicable).

Tests run in this way are not run in parallel.

Note, OpenMP threads are not accounted here, so the total number of
cores used is $(nproc) * NUM_OMP_THREADS

* New test tag `NoConcurrent` and new fixture `TempDirFixture`

Since tests can now be run in parallel, we need to tag tests that cannot
be run concurrently (e.g., because they used shared resources).

A common use case of `NoConcurrent` is for file-system operations.
However, this is a better way to deal with this problem: ensuring
isolation and clean up of tests. To help with this, I added a new
fixture, `TempDirFixture`, that automatically creates and destroy
testing directories.

* script to compute coverage

`measure-test-coverage` is a new script that uses the ctest capabilities
and lcov to distill coverage information and produce html reports.

`measure-test-coverage` can be used to:
1. produce the coverage.info objects by running the tests
2. merge different coverage.info objects
3. produce the html report from the coverage.info files

When run locally, `measure-test-coverage report` will probably cover
most of the use cases. The need to handle different coverage.info
objects might become relevant in the future for CI (if test coverage
information is split across different workers)

* Coverage is now computed as part of CI:

With all the previous advancements, we can now compute coverage as part
of CI and produce a report. The report can be downloaded from the job
page on GitHub actions.