Update coaxial.jl
This commit adds some documentation to coaxial.jl, describing what it
does and how to use it. It also updates the script to be more flexible
with respect to providing a path where `palace` lives.

Running the script I realized that PyPlot no longer seem to work on
Julia 1.12:
```
Hint: Add an appropriate `invokelatest` around the access to this binding.
To make this warning an error, and hence obtain a stack trace, use `julia --depwarn=error`.
┌ Warning: _display is not defined for this backend.
└ @ Plots ~/.julia/packages/Plots/8ZnR3/src/output.jl:218
```
(and the resulting plot is empty)

So, I updated the script to not use PyPlot and remove it as dependency
in the project file.