Fix spack recipe for +superlu/+strumpack with cuda (#570)
* Fix spack recipe for +superlu/+strumpack with cuda

The Palace spack recipe was forcing `strumpack` to be `~cuda` regardless
of the palace variant, so that it was not possible to install
`palace+strumpack+cuda` because of conflicting variants.

Superlu was also not correctly handled.

* Spack: remove forced dependencies on ~cuda/~rocm

There is no strong reason to require that certain dependencies are
compiled without GPU support when Palace is compiled without GPU
support.