Mention Spack Prerequisites (#579)
I investigated Palace's compatibility with WSL as part of #572.

I found that Spack requires certain packages to be present, but does not
advertise this in its installation process. One of the missing packages
is `bzip2`, leading to errors like
```
==> No binary for xz-5.6.3-nkxqqpyo4qf26cpadny6oqsddq6vkdn7 found: installing from source
==> Installing xz-5.6.3-nkxqqpyo4qf26cpadny6oqsddq6vkdn7 [21/70]
==> Fetching https://mirror.spack.io/_source-cache/archive/a9/a95a49147b2dbb5487517acc0adcd77f9c2032cf00664eeae352405357d14a6c.tar.bz2
    [100%]    1.73 MB @   58.4 MB/s
tar (child): lbzip2: Cannot exec: No such file or directory
tar (child): Error is not recoverable: exiting now
/usr/bin/tar: Child returned status 2
/usr/bin/tar: Error is not recoverable: exiting now
```
which are not immediately interpretable as a missing required
dependency.

This commit updates the documentation to draw attention to the
prerequisites.