Add GPU, large ARM runners, and setup-palace-ci
This commit runs the spack workflow on larger/GPU-accelerated runners.

I added a full test matrix, but I could not get everything to work. In
particular, the static builds have several linking errors, which
indicates some issue in the build system.

I tried quick fixes, but I could not get the builds to work, so I am
leaving this for future work.