Compile MFEM externally
This commit puts MFEM on the same footing as the other dependencies.
Instead of being unconditionally built, it is now build when building
external dependencies. This allowed me to add MFEM as a dependency
managed spack.

In doing this, I found a new bug in spack:
https://github.com/spack/spack/issues/51505

So, I changed our approach to testing the spack package. Instead of
adding a local repo, I overwrite palace in the `builtin` repo.